Update Export to use empty
There are multiple locations where data is loosely tested against != '' or != '0'.  Updated the export file to instead test against empty with consideration of all values that would identify the object as empty being acceptable and possible for that location.  Tests like !== false as used when searching text has not been touched/adjusted.