Co-locate EP4_DB_FILTER_KEY import information
Created 2 new variables to support removing multiple occurrences
of testing for the primary key by moving all of the assignments
just inside of the first record import test/check and before
usage of other "external" files.  This way the information is
available, set one time and used where/as necessary.

The variables are:
$chosen_key (the field to be used as the primary key prefixed with v_)
$chosen_key_sql (the primary key as compared to the value it is
                 set when replaced with $db->bindVars accessing
                 the table aliased with a p.)
$chosen_key_sql_limit (essentially the primary_key with a limit
                       of 1 applied.)

With movement of chosen key related information into the main
part of the import file, have removed duplicate code from
sub-sections to simplify processing and be more consistent.