Implementation of additional sanitization affects text with NULL
A string field that had NULL in it would be changed to a value of
null regardless of what else was in the string unless the
type conversion of stringIgnoreNull was used.  Problem is that
the type conversion is not applied across all of ZC therefore
a method of testing for its presence was needed that would work
across ZC versions.  That test and its use have been implemented
by this commit.

Fixes #39