Send data to browser to keep-alive and add @set_time_limit directive
After all of these years of reports that EP4 times out and
the inability to import/export a large file.  Finally a potential
solution has been considered that may just solve the issue in
the most generic of cases.  There still are some outliers or
system conditions that may prevent successful use of this
additional feature; however, this should be an observable
improvement.

Where files or large number of records are being processed have
added in the set_time_limit directive to allow PHP to continue
processing even if the web server portion of the operation
times out.  The web portion could be addressed by incorporating
something that would update the screen.

This feature had been a consideration a few years ago; however,
on the one system where it seemed like it would make a difference
it did not work and so was removed from the code.  Continued
review has suggested that EP4 would benefit from using it.