Initial strict debugging.
This commit is the result of running both import and export
of files through EP4 using the standard file export and
little to no modification of the file for import.  There
may still be additional modifications to incorporate to prevent
additional notifications being generated when working with
partials of a file.