Incorporate option to export date in yymmdd format
When several export files are generated, the default export
filename format has the date listed as YYYYMonDD.  Depending
on the language used, the sort order of these files is
primarily within a given month, but a long list of records
is not in chronological order.  By exporting the date in a
numerical format YYYYMMDD then when sorted by filename, the
oldest record is at top, the newest is at bottom.

Added the configuration key information to install/add
the option for export files to include the date format
of YYYYmmdd to support improved display sorting.


Alternatively, the display of records could be by date order
and not filename; however in that situation, an older file
that has recently been modified will show at the top...