New category_id based on auto_number
Update the category_id to be based on the next auto_number instead of the next highest number that is available in the categories table.  Ie. if category 100 was the last added and it was then deleted, the code previously would use 100 next; however, as a relational database the thought is that it would be the next number not used and therefore should be 101.  The products_id is determined this way.