Prevent products name and others from being empty
While incorporating an array method to carry information across the code, A condition was introduced where array variables were set to an empty string instead of not existing which would cause a problem later in processing when expecting that if the variable was set that it would be used for data assignment.  This occurred for language related variables such as `products_name`, `products_short_desc`, `products_description`, and `products_url`. A reason that the empty string option is not good is because then the array variable could be set as an empty string for two reasons, one it is a new product where the column was not included and two it was a new product where the column was included and the value was an empty string.