Assign a default value for categories_description
When generating a new category using EP4 while adding a new product, the category name is inserted; however, there has not been a description entered yet and therefore there was no data available to import.  A default ZC install does not assign a default value for the categories_description field within the categories_description table.  

This adds a default value; however, the database should be first queried to identify the existence of a default value.  If a default value is there, then either use the default value or don't include the categories_description when inserting that record.
This relates to #47.