Incorporate $chosen_key into attribute import
When redesigning the use and designation of primary keys within EP4, the `$chosen_key` related information was identified to be prefixed with 'v_'.  This prefix is currently needed throughout the remainder of the code; however, could either be eliminated or further incorporated.  This is a subtle fix to that issue until a more permanent solution is incorporated.  This detects the presence of the prefix of 'v_' for the `$chosen_key` and removes it to be able to perform the substitution in the `$chosen_key_sql` statement with the statement left as written.