Corrects the data type of data being assigned.
There were pieces of data that were set to sanitize; however,
sanitizing 0 to a string causes an entry string instead of like
when sanitizing 0 to an integer where it is stored as a value
of zero.  This corrects the conversions that have been
identified to be incorporated into the next version.