Add product code override for categories and full product.
The default is set for the language code to override the
language id.  What this means is that if a field supports
multiple languages, if the file contains the field using its
language id (1, 2, 3, etc...) and its language
code (en, es, ge, etc...) then with language code overriding
the values used with the language code will be what are stored.

At this time only language_code and language_id are incorporated
for pulling data from the file(s).