Do not loop on file if the primary key is not in the file.
The first action inside of this loop is to pull data from the
database based off of the primary key that has been selected.
If the primary key is not in the import file, then the associated
query of the database will not provide appropriate data and
multiple notices will be generated.  This prevents attempting
to process the file if there is no primary key in it.

Fixes #50.