Update filelayout to use new method of discount quantities
As identified by @swguy at: https://www.zen-cart.com/showthread.php?190417-EasyPopulate-4-0-Support-Thread&p=1360571#post1360571 there *IS* a better way to determine the number of quantity discounts to consider exporting rather than manually entering the number.  This change supports the auto-determination if the value is left blank or set to 0 with a default of 3.  The value of 3 was previously used as a starting point.

The reason for a default is that in an empty store or one that has no quantity discounts, an export of the data for quantity discounts would provide no columns.  To ensure a level of understanding of the meaning of the suffix for v_discount_qty_X and v_discount_price_X a minimum number (`$num_discounts`) of quantity discount columns will be exported regardless the actual number in the database. The minimum number is set to 3 as was historically a part of EP4, but could be modified in the code to some other number.