Default install qty discount columns to auto-detect.
For new installs or where the settings have been removed from the database, set the default determination of the maximum number of quantity discounts to auto-determine the number of columns.  The code has been modified to provide a minimum quantity based on the historical default (`$num_discounts`); however, code editors may change that value from 3.