Move $messageStack notices out of import file to streamline cron.
Variables that cause display output when a cron job is executed
require additional handling.  Removing the `$messageStack` out
of the import file allows the cron job to complete and should not
cause any myDebug log files to be created.  The data is still
collected to be provided, but is displayed at completion of the
import.