Refactor and add multiple language field designation import.
Adds the ability to import the category meta file using language extensions
of _1 or _en for the language that is assigned the language_id of 1 which
may also be english.  Provides for ability to control which field listing
takes priority if both are provided, otherwise it assigns the information
to the language assigneed/identified.