Correct import of basic attributes
In attempting to clearly identify the expected value indicating that the
retrieved data is to stop, the comparison against `false` for both
`mysql_fetch_array` and `mysqli_fetch_array` was found to be wrong.
The `mysqli_fetch_array` function now returns `null` instead of `false` as
it did in the older style.

Further, in review of the fields used in one of the arrays to support
expanding capability for optional primary keys, the key used
was not part of the associaated dataset. Updated to use the database version
of the primary key.

Further, one of the queries that have historically been used/reused has
been updated to be an independent query and not replace a previous query.