Update sort order based on previous determination
In an update situation, the option value sort order was being populated by
the numeric position of the option value starting at a value of 1.

Meanwhile new attributes were beginning at a starting value and incrementing
by whole numbers of 10.  This modifies the update process to use the same
incremental number as a new option value so that option value sort orders
will be incremenbed by whole numbers of 10 offering "space" for manual
updates between the uploaded option value.