Incorporate handling for products_options_type.
Offers a default products_options_type for imports that do not include
the field.  Further the products_options_type within the file is based on
database settings/value instead of hard coded numerical values.

Have also incorporated the ability to modify the products_options_type
for an option name. This update of type will be based on the products_options_name
and will be applied to all option names in the language supported.