Correct notice/warning messages
The define may not have been executed through normal operation; however,
in reviewing the code, the absence of single quotes was noticed and is
corrected here which would have generated a myDEBUG-adm related log.

The build up of the $messageStack information was identified as an issue
by more strict php controls where the php end function needs to use a
variable not the result of another command.  The information to be fed
to the end command (retrieval of the last element of the array) was
therefore placed into a separate variable to be processed.