Expand support for alternate primary keys.
This is intended to offer notifier access to use/consider an alternate
primary key than the default version(s) already incorporated.  Generally
speaking the primary key would be based on the products table and that is
still the expectation though alternate concepts may be developed/suggested.

An applicable observer would be needed to address/handle all of the cases/conditions
where the data is to be modified.  Such an example is provided in the forum
as part developing this.