Revise the binding to allow NULL to be retained.
Over time, a string that contained the word NULL in it was made to actually
become null unless a different method of conversion was used that allowed
the word null to remain in its value. This had been done more wholesale
previously in other files and there may be other similar modifications available
for this file, but this seems to coincide with the other change(s) adjacent
to these two lines.