Start adding capability for multiple specials on product
Default ZC operation supports a single special to be applied to each product and that the `specials` table contain only one special per product. It appears that the largest reason for that is to prevent/mitigate the issue of two specials having overlapping dates.  This change initiates operations to support insertion of multiple specials for a product and also begins providing a path of how to mitigate inserting multiple specials for a single product including the potential of the result of overlapping dates.

There is the possibility that an additional function be "created" based on the version of Zen Cart and existence of `zen_datetime_overlap` that became available in Zen Cart's `includes/functions/functions_general.php`