Swap the message of install and missing folder.
While the installation process can lead to the dismissal of the missing
installation folder, the notification of needing to define an installation
folder is still considered useful information at the first visit to an
uninstalled version of EP4 and therefore remains displayed. The sequence
though has been reversed so that there is the possibility of the message
going away if the "default" folder is present, otherwise the message will
be updated to reflect the name of the folder sought.