Fully support import by language_code.
EP4 had been built to import language centric items using the language_id
which is the number assigned to the language by the database.  While this
was made functional, it adds some complexity to the user trying to import
information especially in a multi-language database or more so to multiple
databases.  One database may use the number 2 for another language but use
the number 3 for that same language elsewhere.  While this effort has been
in the making for revision for some time, it is expected that this commit
will make the code full for supporting the language_code (basically 2 letter
designation for the language) as a language identifier.

THere are now 4 settings, language_id, language_code, language_id_only and
language_code_only.  The first two are "suggestions" where the software
will use whichever is selected if it exists otherwise the value of the
other option type will be used.  If either of the last two options are
set then the field of that type will be the only type field to be used.

Example: if import a file that has both v_products_name_1 and v_products_name_en,
then if the setting is language_code or language_code_only then
v_products_name_en will be used.
If the file only had the field v_products_name_1, and the setting is language_code
the v_products_name_1 will be imported, but if it is set to language_code_only
then the column v_products_name_1 will be skipped.

This feature is offered to support the historical use, a transitional use,
and a multi-faceted use.  Ideally the setting would be made to import one
field type only and all files would either be imported only with that field
type or that edits will be made only to fields that have that designation.

I expect there to be an increase in support requests because of the additional
capability.  I did not want to eliminate the use of the language_id as
there have been so many users that have shown interest in this plugin and
appear to be using it that completely dropping the use of the language_id
may be significantly problematic to those businesses. While I take no
liability for issues that they may experience, I do not want to directly
cause issues that could have been avoided.