Correct default install and minor refactor
Decided to return initial operation to use of the language_id.

EP4 has been known to use it, so might as well bank on it. The capability
to use the language_code is present, the ability for a file to have both
and for the software to "sort it out" is there, but the community has
expectations and have come to know this uses the language_id, might as
well ensure that it is supported throughout.