Reduce notifications made on new store pre-database install
Disables attempts to generate EP4 debug logs if EP4 is not installed.
Even though the logic supports not having the debug logging configuration
installed, that is a factor associated with previous versions of EP4 and
options made available by other admins. This change though prevents
attempting to create debug log files through the EP4 logging system because
without the temporary folder identified, there is "no place" to store the
logs and further there is not really supposed to be much if anything
performed in this condition.