Correct export of SBA Stock files
In certain initial rows of data, a condition might have existed to
reduce the number of columns in that row by 1. This would cause an
unknown CSV delimiter messaage associated with import.

This ensures that the field v_SBA_tracked is present and where
necessary that v_customid is also included/cleared.