Update default import overrides
The database setting for the default is being modified to
default to language_id_only on new upgrade, if the options
were previously present for testing then there is no need
to modify the default setting. If the database setting isn't
made, then the files are to use the language_id first and then
the language_code if the associated _id field doesn't exist.

This is in part to stay with the original settings for EP4
where the language_id was all that was used.  Now the language_code
can also be used, but data may be out-of-date upon import.