Robustly add attribute export by upper filters.
This added indication to the category dropdown list to show category with
product and those that have product with attributes.

Includes ajax updates to switch between showing both product and product
with attributes to just showing product or whatever display style is
coded.

Add Ajax list update code.

Incorporated flags to update showing
1) product that have attributes and
2) categories that have product.

This includes defining a constant `TOPMOST_CATEGORY_PARENT_ID` that
may not get defined as a result of core Zen Cart design/implementation.

Unfortunately this slightly broke the possibility of fully installing via
the plugin manager for Zen Cart 1.5.7 because there is a catalog side file
involved even for admin code.

Add/update AJAX code and ensure load of default language

In the event that an alternate language file is not present at least load
the english file. Standard ZC functions to do this have not yet been
incorporated.