Update Queries to ANSI 92 standard
Although notifiers exist in the code to modify queries, with the queries
not in ANSI 92 form, joins possibly did not align as expected/desired.

By this update, improvements are available to more readily adapt the
queries to desired questions. This includes updating table prefixes
to align with majority of such in Zen Cart and to ensure internal alignment.