Modify Ajax loader path
Not a normal situation; however, if the software has been installed as a
plugin, then the files are removed from the plugin directory and instead
placed in the catalog, then the load path is attempted/expected to be
through the file as in the admin directory. If it is not there, then
the additional feature is simply not loaded/available.