Revise Category insert because of mySQL 8.x
A change to default settings in mySQL 8.x that are not always able to be
overridden by an end user causes database table configured data to remain
cached for an extended period. As a result, attempts to insert new
categories were failing because the category id already existed.

Historically category addition was done based on the highest value retrieved
from the database which may not reflect the highest value ever used. Further,
the value needs to be dynamic as other users may be adding a category at the
same time and cause some sort of clash if not handled in the correct
sequence. Now though, this insert will fail if the categories table has
lost its auto_increment condition (which was not a problem per se with
previous methods of determination).

Fixes #67