Changing products_type search to search for name of handler
Currently the v_products_type field does a search by the product's type_id as defined in the products table. This code will change it so that instead of looking for the product's `type_id`, it will search for the product's `type_handler`. If the search fails, it will assume the default product type of 1.